+++ src/main/java/com/api/MoriMagazineAPI/Controller/ProdutoControllerWeb.java
@@ -2,7 +2,6 @@ package com.api.MoriMagazineAPI.controller;
 
 import com.api.MoriMagazineAPI.data.ProdutoEntity;
 import com.api.MoriMagazineAPI.service.ProdutoService;
-
 import jakarta.validation.Valid;
 import java.math.BigDecimal;
 import java.util.List;
@@ -26,8 +25,8 @@ public class ProdutoControllerWeb {
     }
 
     @GetMapping("/deletar/{id}")
-    public String mostrarFormularioExclusao(@PathVariable(value = "id") Long id, Model model) {
-        ProdutoEntity produto = produtoService.getProdutoId(id);
+    public String mostrarFormularioExclusao(@PathVariable(value = "id") Integer id, Model model) {
+        ProdutoEntity produto = produtoService.getProdutoId(id); // Ajustado para o método correto
         if (produto == null) {
             return "redirect:/produto/listar";
         }
@@ -36,7 +35,7 @@ public class ProdutoControllerWeb {
     }
 
     @PostMapping("/deletar/{id}")
-    public String deletarProduto(@PathVariable(value = "id") Long id) {
+    public String deletarProduto(@PathVariable(value = "id") Integer id) {
         produtoService.deletarProduto(id);
         return "redirect:/produto/listar";
     }
@@ -52,7 +51,6 @@ public class ProdutoControllerWeb {
         if (result.hasErrors()) {
             return "inserirProduto";
         }
-
         try {
             String precoFormatado = produto.getPreco().toString().replaceAll("[^\\d.,]", "");
             precoFormatado = precoFormatado.replace(",", ".");
@@ -61,11 +59,10 @@ public class ProdutoControllerWeb {
         } catch (NumberFormatException e) {
             e.printStackTrace();
         }
-
         if (produto.getId() == null) {
             produtoService.criarProduto(produto);
         } else {
-            produtoService.atualizarProduto(produto.getId(), produto);
+            produtoService.atualizarProduto(produto.getId(), produto); // Ajustado para passar o ID
         }
         return "redirect:/produto/listar";
     }
@@ -78,18 +75,17 @@ public class ProdutoControllerWeb {
     }
 
     @GetMapping("/atualizarForm/{id}")
-    public String mostrarFormularioAtualizacao(@PathVariable(value = "id") Long id, Model model) {
-        ProdutoEntity produto = produtoService.getProdutoId(id);
+    public String mostrarFormularioAtualizacao(@PathVariable(value = "id") Integer id, Model model) {
+        ProdutoEntity produto = produtoService.getProdutoId(id); // Ajustado para o método correto
         model.addAttribute("produto", produto);
         return "atualizarProduto";
     }
 
     @PostMapping("/atualizar/{id}")
-    public String atualizarProduto(@PathVariable(value = "id") Long id, @Valid @ModelAttribute("produto") ProdutoEntity produto, BindingResult result) {
+    public String atualizarProduto(@PathVariable(value = "id") Integer id, @Valid @ModelAttribute("produto") ProdutoEntity produto, BindingResult result) {
         if (result.hasErrors()) {
             return "atualizarProduto";
         }
-
         try {
             String precoFormatado = produto.getPreco().toString().replaceAll("[^\\d.,]", "");
             precoFormatado = precoFormatado.replace(",", ".");
@@ -98,14 +94,13 @@ public class ProdutoControllerWeb {
         } catch (NumberFormatException e) {
             e.printStackTrace();
         }
-
-        produtoService.atualizarProduto(id, produto);
+        produtoService.atualizarProduto(id, produto); // Ajustado para passar o ID
         return "redirect:/produto/listar";
     }
 
     @GetMapping("/buscarPorId/{id}")
-    public String buscarProdutoPorId(@PathVariable("id") Long id, Model model) {
-        ProdutoEntity produto = produtoService.getProdutoId(id);
+    public String buscarProdutoPorId(@PathVariable("id") Integer id, Model model) {
+        ProdutoEntity produto = produtoService.getProdutoId(id); // Ajustado para o método correto
         if (produto == null) {
             return "redirect:/produto/listar";
         }
